Move TFilters type export to commit-graph-sidebar
